entsoe-streaming/requirements.txt
kafka-python
confluent_kafka
requests
avro
faust
fastavro
python-dotenv